requirements.txt
python-telegram-bot[job-queue]
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic-settings
alembic
python-dotenv
httpx
greenlet